Moved representations and request parameters from Identity Module
