+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/representation/UserRepresentation.java
@@ -9,15 +9,16 @@
  */
 package org.oscm.rest.common.representation;
 
+import org.oscm.internal.types.enumtypes.OrganizationRoleType;
 import org.oscm.internal.types.enumtypes.Salutation;
+import org.oscm.internal.types.enumtypes.UserAccountStatus;
+import org.oscm.internal.types.enumtypes.UserRoleType;
 import org.oscm.internal.vo.VOUserDetails;
-import org.oscm.rest.common.representation.Representation;
 
 import javax.ws.rs.WebApplicationException;
+import java.util.*;
 
 public class UserRepresentation extends Representation {
-
-  private transient VOUserDetails ud;
 
   private String email;
   private String firstName;
@@ -28,54 +29,66 @@ public class UserRepresentation extends Representation {
   private String locale;
   private Salutation salutation;
   private String realmUserId;
+  private boolean remoteLdapActive;
   private String organizationId;
   private String userId;
+  private UserAccountStatus status;
+  private Set<OrganizationRoleType> organizationRoles = new HashSet<OrganizationRoleType>();
+  private Set<UserRoleType> userRoles = new HashSet<UserRoleType>();
+
+  private transient VOUserDetails vo;
 
   public UserRepresentation() {
     this(new VOUserDetails());
   }
 
   public UserRepresentation(VOUserDetails details) {
-    ud = details;
+    vo = details;
   }
 
   @Override
-  public void validateContent() throws WebApplicationException {
-    // nothing right now
-  }
+  public void validateContent() throws WebApplicationException {}
 
   @Override
   public void update() {
-    ud.setAdditionalName(getAdditionalName());
-    ud.setAddress(getAddress());
-    ud.setEMail(getEmail());
-    ud.setFirstName(getFirstName());
-    ud.setKey(convertIdToKey());
-    ud.setLastName(getLastName());
-    ud.setLocale(getLocale());
-    ud.setOrganizationId(getOrganizationId());
-    ud.setPhone(getPhone());
-    ud.setRealmUserId(getRealmUserId());
-    ud.setSalutation(getSalutation());
-    ud.setUserId(getUserId());
-    ud.setVersion(convertETagToVersion());
+    vo.setAdditionalName(getAdditionalName());
+    vo.setAddress(getAddress());
+    vo.setEMail(getEmail());
+    vo.setFirstName(getFirstName());
+    vo.setKey(convertIdToKey());
+    vo.setLastName(getLastName());
+    vo.setLocale(getLocale());
+    vo.setOrganizationId(getOrganizationId());
+    vo.setOrganizationRoles(getOrganizationRoles());
+    vo.setPhone(getPhone());
+    vo.setRealmUserId(getRealmUserId());
+    vo.setRemoteLdapActive(isRemoteLdapActive());
+    vo.setSalutation(getSalutation());
+    vo.setStatus(getStatus());
+    vo.setUserId(getUserId());
+    vo.setUserRoles(getUserRoles());
+    vo.setVersion(convertETagToVersion());
   }
 
   @Override
   public void convert() {
-    setAdditionalName(ud.getAdditionalName());
-    setAddress(ud.getAddress());
-    setEmail(ud.getEMail());
-    setFirstName(ud.getFirstName());
-    setId(Long.valueOf(ud.getKey()));
-    setLastName(ud.getLastName());
-    setLocale(ud.getLocale());
-    setOrganizationId(ud.getOrganizationId());
-    setPhone(ud.getPhone());
-    setRealmUserId(ud.getRealmUserId());
-    setSalutation(ud.getSalutation());
-    setETag(Long.valueOf(ud.getVersion()));
-    setUserId(ud.getUserId());
+    setAdditionalName(vo.getAdditionalName());
+    setAddress(vo.getAddress());
+    setEmail(vo.getEMail());
+    setFirstName(vo.getFirstName());
+    setId(Long.valueOf(vo.getKey()));
+    setLastName(vo.getLastName());
+    setLocale(vo.getLocale());
+    setOrganizationId(vo.getOrganizationId());
+    setOrganizationRoles(vo.getOrganizationRoles());
+    setPhone(vo.getPhone());
+    setRealmUserId(vo.getRealmUserId());
+    setRemoteLdapActive(vo.isRemoteLdapActive());
+    setSalutation(vo.getSalutation());
+    setStatus(vo.getStatus());
+    setETag(Long.valueOf(vo.getVersion()));
+    setUserId(vo.getUserId());
+    setUserRoles(vo.getUserRoles());
   }
 
   public String getOrganizationId() {
@@ -102,8 +115,24 @@ public class UserRepresentation extends Representation {
     this.userId = userId;
   }
 
+  public void setStatus(UserAccountStatus status) {
+    this.status = status;
+  }
+
+  public UserAccountStatus getStatus() {
+    return status;
+  }
+
+  public Set<UserRoleType> getUserRoles() {
+    return userRoles;
+  }
+
   public String getAddress() {
     return address;
+  }
+
+  public void setUserRoles(Set<UserRoleType> userRoles) {
+    this.userRoles = userRoles;
   }
 
   public void setAddress(String address) {
@@ -116,6 +145,14 @@ public class UserRepresentation extends Representation {
 
   public void setPhone(String phone) {
     this.phone = phone;
+  }
+
+  public Set<OrganizationRoleType> getOrganizationRoles() {
+    return organizationRoles;
+  }
+
+  public void setOrganizationRoles(Set<OrganizationRoleType> organizationRoles) {
+    this.organizationRoles = organizationRoles;
   }
 
   public String getEmail() {
@@ -166,7 +203,23 @@ public class UserRepresentation extends Representation {
     this.salutation = salutation;
   }
 
+  public boolean isRemoteLdapActive() {
+    return remoteLdapActive;
+  }
+
+  public void setRemoteLdapActive(boolean remoteLdapActive) {
+    this.remoteLdapActive = remoteLdapActive;
+  }
+
   public VOUserDetails getVO() {
-    return ud;
+    return vo;
+  }
+
+  public static final Collection<UserRepresentation> convert(List<VOUserDetails> list) {
+    Collection<UserRepresentation> result = new ArrayList<UserRepresentation>();
+    for (VOUserDetails vo : list) {
+      result.add(new UserRepresentation(vo));
+    }
+    return result;
   }
 }
