+++ oscm-rest-api-common/src/main/java/org/oscm/rest/common/requestparameters/UserParameters.java
@@ -7,7 +7,7 @@
  *
  * <p>*****************************************************************************
  */
-package org.oscm.rest.identity;
+package org.oscm.rest.common.requestparameters;
 
 import org.oscm.rest.common.CommonParams;
 import org.oscm.rest.common.requestparameters.RequestParameters;
