+++ oscm-rest-api-identity/src/main/java/org/oscm/rest/identity/RolesResource.java
@@ -14,7 +14,8 @@ import lombok.Setter;
 import org.oscm.rest.common.CommonParams;
 import org.oscm.rest.common.RestResource;
 import org.oscm.rest.common.Since;
-import org.oscm.rest.identity.data.RolesRepresentation;
+import org.oscm.rest.common.requestparameters.UserParameters;
+import org.oscm.rest.common.representation.RolesRepresentation;
 
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
