+++ oscm-rest-api-identity/src/test/java/org/oscm/rest/identity/UserBackendTest.java
@@ -22,9 +22,10 @@ import org.oscm.internal.intf.IdentityService;
 import org.oscm.internal.vo.VOUserDetails;
 import org.oscm.rest.common.representation.RepresentationCollection;
 import org.oscm.rest.common.SampleTestDataUtility;
-import org.oscm.rest.identity.data.OnBehalfUserRepresentation;
-import org.oscm.rest.identity.data.RolesRepresentation;
-import org.oscm.rest.identity.data.UserRepresentation;
+import org.oscm.rest.common.requestparameters.UserParameters;
+import org.oscm.rest.common.representation.OnBehalfUserRepresentation;
+import org.oscm.rest.common.representation.RolesRepresentation;
+import org.oscm.rest.common.representation.UserRepresentation;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
