+++ oscm-rest-api-identity/src/test/java/org/oscm/rest/identity/data/RolesRepresentationTest.java
@@ -16,6 +16,7 @@ import org.oscm.internal.vo.BaseVO;
 import org.oscm.internal.vo.VOUser;
 import org.oscm.internal.vo.VOUserDetails;
 import org.oscm.rest.common.representation.Representation;
+import org.oscm.rest.common.representation.RolesRepresentation;
 
 import static org.assertj.core.api.Assertions.assertThat;
